LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/layer-counter-style-override-expected.txt


FAIL @counter-style unlayered overrides layered assert_equals: expected "31.21875px" but got "0px"
FAIL @counter-style override between layers assert_equals: expected "31.21875px" but got "6px"
FAIL @counter-style override update with appended sheet 1 assert_equals: expected "31.21875px" but got "6px"
FAIL @counter-style override update with appended sheet 2 assert_equals: expected "31.21875px" but got "6px"
